NLP/notification_service/requirements.txt
sib-api-v3-sdk
python-dotenv
mailjet-rest